backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1
sqlparse==0.4.4
bcrypt==4.1.2
pyjwt==2.8.0
